requirements.txt
Django~=5.0.3
pytest~=8.2.2
Faker~=12.0.1
mixer~=7.2.2
django-debug-toolbar~=4.4.5
pip~=24.0
attrs~=23.1.0
pillow~=10.3.0
Jinja2~=3.1.2
djangorestframework~=3.15.2